requirements.txt
matplotlib
black
flake8
isort
loguru
mkdocs
pip
python-dotenv
setuptools
typer
opencv-python-headless
tqdm 
pillow
deeplabcut[tf]
-e .
